+++ adapters/Line/LineAdapter/GADMediationAdapterLineInterstitialAdLoader.m
@@ -137,7 +137,7 @@
 - (void)presentFromViewController:(nonnull UIViewController *)viewController {
   GADMediationAdapterLineLog(@"FiveAd SDK will present the interstitial ad.");
   [_interstitialAdEventDelegate willPresentFullScreenView];
-  [_interstitialAd show];
+  [_interstitialAd showWithViewController:viewController];
 }
 
 #pragma mark - FADLoadDelegate (for waterfall interstitial ad)